Rename MutableArray to ArrayImitatorBuilder
And reorder all methods by name ASC
